Refactor branch equation term
